examples/python/requirements.txt
flask==3.0.0
anthropic==0.18.1
python-dotenv==1.0.0
requests==2.31.0
